Delete pictures
- Created a Helper class within the src directory
- Static method to delete files with an option boolean to show the user a confirmation message
