+++ PicturebotGUI/PicturebotGUI.csproj
@@ -60,6 +60,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="src\Helper\Helper.cs" />
     <Compile Include="src\POCO\Config.cs" />
     <Compile Include="Extension.cs" />
     <Compile Include="Flow.cs" />
@@ -87,7 +88,6 @@
     <Compile Include="src\Froms\FormShoot.Designer.cs">
       <DependentUpon>FormShoot.cs</DependentUpon>
     </Compile>
-    <Compile Include="Helper.cs" />
     <Compile Include="Picture.cs" />
     <Compile Include="src\POCO\Canvas.cs" />
     <Compile Include="Program.cs" />
@@ -129,5 +129,6 @@
   <ItemGroup>
     <None Include="App.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
